requirements.txt
absl-py==0.8.0
astor==0.8.0
beautifulsoup4==4.6.3
blessed==1.15.0
botocore==1.12.233
cached-property==1.5.1
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
Django==2.2.5
django-cors-headers==2.5.2
djangorestframework==3.10.2
dnspython==1.16.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
gast==0.3.2
google-pasta==0.1.7
grpcio==1.24.0
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
jmespath==0.9.4
joblib==0.13.2
jsonschema==2.6.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
numpy==1.17.2
pandas==0.25.1
pathspec==0.5.9
protobuf==3.9.2
pymongo==3.9.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.20.1
scikit-learn==0.21.3
scipy==1.3.1
semantic-version==2.5.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.1
uWSGI==2.0.18
virtualenv==16.7.5
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
